Estimator/requirements.txt
kafka-python
numpy
scipy
pytest
